Extract out `TargetMap` and `WsConnection` from `ApiServer`.
...and rename `ApiServer` -> `WsConnection`.
