+++ local-modules/api-server/main.js
@@ -2,6 +2,8 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import ApiServer from './ApiServer';
+import Target from './Target';
+import TargetMap from './TargetMap';
+import WsConnection from './WsConnection';
 
-export { ApiServer };
+export { Target, TargetMap, WsConnection };
